requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
ollama
pydantic
pydantic-settings
sentence-transformers
